Добавлена система роста растений
Добавлена возможность выращивать растения, а так же смена дня и ночи.

+++ Assets/Scripts/ScriptableObjects/Item.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public abstract class Item : ScriptableObject
 {
-    public float MaxStack;
+    public int MaxStack;
     public Sprite ItemSprite;
     public UsableItem ItemUsage;
     public RuntimeAnimatorController Animator;
